pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xlerobot-personality"
version = "0.1.0"
description = "Starter system for pan/tilt tracking + scene interaction on XLeRobot head"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "opencv-python>=4.8",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
piper = ["piper-tts>=1.4.1"]
yolo = ["ultralytics>=8.0.0"]
dev = ["pytest>=8.0"]

[project.scripts]
xlerobot-personality = "xlerobot_personality.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
